+++ app/server.js
@@ -1,33 +1,25 @@
-// Server.js - This file is the initial starting point for the Node/Express server.
-//
-// ******************************************************************************
-// *** Dependencies
-// =============================================================
-var express = require("express");
-var bodyParser = require("body-parser");
+// Dependencies
+const bodyParser = require("body-parser");
+const express = require("express");
+const handlebars = require("express-handlebars");
+const path = require("path");
+const routes = require("./controllers/articleRoutes");
 
-// Sets up the Express App
-// =============================================================
-var app = express();
-var PORT = process.env.PORT || 8080;
-
-// Requiring our models for syncing
-// var db = require("./models");
+const app = express();
+const PORT = process.env.PORT || 8080;
 
 // Sets up the Express app to handle data parsing
-
-// parse application/x-www-form-urlencoded
 app.use(bodyParser.urlencoded({ extended: true }));
-// parse application/json
 app.use(bodyParser.json());
 
-// Static directory
-app.use(express.static("views"));
+// Set Handlebars for templates
+app.engine("handlebars", handlebars({ defaultLayout: "main" }));
+app.set("view engine", "handlebars");
 
-// Routes
-// =============================================================
-require("./routes/approutes.js")(app);
+// Use static folder to serve pages
+app.use(express.static(path.join(__dirname, "public")));
+app.use(routes);
 
-app.listen(PORT, function() {
-    console.log("App listening on PORT " + PORT);
-  });
+app.listen(PORT, () => {
+    console.log(`Server started, listening on PORT ${PORT}...`);
+});
