--- db/schema.sql
@@ -1,3 +1,0 @@
-DROP DATABASE IF EXISTS harkup_db;
-CREATE DATABASE harkup_db;
-USE DATABASE harkup_db;
