--- db/seeds.sql
@@ -1,1 +1,0 @@
--- Some seed data will go in this file
